+++ grip_api/src/grip_api/config_widgets/code_editors.py
@@ -17,199 +17,10 @@
 import re
 import copy
 from PyQt5 import Qsci
-from PyQt5.QtGui import QColor
-from PyQt5.QtCore import pyqtSignal, QTimer, Qt
+from grip_api.src.grip_api.abstract_widgets.base_text_editor import BaseCodeEditor
 
 
-class GenericCodeEditor(Qsci.QsciScintilla):
-
-    """
-        QScintilla-based widget allowing to create a generic code editor
-    """
-    # Signal sent when the the parsed content changes
-    contentIsModified = pyqtSignal(bool)
-
-    def __init__(self, parent=None):
-        """
-            Initialize the class by setting up the editor
-
-            @param parent: parent of the widget
-        """
-        super().__init__(parent)
-        self.is_lexed = False
-        self.init_ui()
-        self.init_backround_markers()
-        self.lexer_ = None
-        self.initial_content = {}
-        # Dictionary that will have as keys the content of the line and as value the index of the line of new dicts
-        self._new_dicts_index = {}
-        # Will contain the index of the lines wrongly formatted
-        self.wrong_format_lines = []
-        # Timer used to check content format and that will be handled by a different thread
-        self.timer = QTimer()
-        # Set it to single shot (i.e. does not run continuously)
-        self.timer.setSingleShot(True)
-        # Once the timer is timing out then starts the check
-        # We do this to avoid having stuff signaled as wrong while editing
-        self.timer.timeout.connect(self.parse_and_format_editor)
-        # Each time a new character is inserted in the editor, restart the timer
-        self.textChanged.connect(self.start_timer)
-
-    @property
-    def is_lexed(self):
-        """
-            Return a boolean stating wheter the editor is lexed or not
-
-            @return: Boolean indicating if the editor is lexed
-        """
-        return self._is_lexed
-
-    @is_lexed.setter
-    def is_lexed(self, boolean_value):
-        """
-            Set the boolean indicating whether the editor is lexed or not
-
-            @param boolean_value: Boolean indicating if the editor is lexed or not
-        """
-        if not isinstance(boolean_value, bool):
-            raise TypeError("The input value for 'is_lexed' must be a boolean")
-        # Assing the input value if everything is all right
-        self._is_lexed = boolean_value
-
-    def init_ui(self):
-        """
-            Initialize the editor
-        """
-        # By default no lexer is set
-        self.setLexer(None)
-        # Set a grayish colour
-        self.empty_color = QColor("#cccccc")
-        self.setPaper(self.empty_color)
-        # Set the tab width to 2 to save space
-        self.setTabWidth(2)
-        # Change tabs to spaces
-        self.setIndentationsUseTabs(False)
-        # Help to visualize the indentation
-        self.setIndentationGuides(True)
-        # Set auto indentation
-        self.setAutoIndent(True)
-        # Cannot be edited by the user
-        self.setReadOnly(True)
-        # Remove some of the standard shortcut embedded in QScintilla
-        commands = self.standardCommands()
-        commands.boundTo(Qt.ControlModifier | Qt.Key_L).setKey(0)
-        commands.boundTo(Qt.ControlModifier | Qt.Key_T).setKey(0)
-
-    def init_backround_markers(self):
-        """
-            Define markers to highlight lines not properly formatted using a red-ish colour
-        """
-        self.markerDefine(Qsci.QsciScintilla.Background, 0)
-        self.setMarkerBackgroundColor(QColor("#40FF0000"), 0)
-
-    def parse_and_format_editor(self):
-        """
-            Run the parser on the editor's content and signal which lines are not well formatted
-        """
-        # Parse the content of the editor
-        self.parse_content()
-        # Make the background of wrongly formatted lines red-ish
-        self.update_background()
-
-    def start_timer(self):
-        """
-            Start the timer that triggers the content's format checking
-        """
-        # The timer would timeout after 600 ms meaning that the check would happend 600ms after the last text edit
-        self.timer.start(600)
-
-    def stop_timer(self):
-        """
-            Stop the timer that triggers the content's format checking. After running this function, the content that
-            will be set to the editor WON'T be automatically parsed and checked.
-        """
-        self.timer.stop()
-
-    def set_text_and_trigger_checks(self, content):
-        """
-            Set the content of the editor and immedialty trigger checks about its validity
-
-            @param content: String to be displayed in the editor
-        """
-        # Stop the timer (so we can trigger checks without having to wait for 600 ms)
-        self.stop_timer()
-        # Set the text and immediately trigger the checks
-        self.setText(content)
-        self.parse_and_format_editor()
-        # Restart the timer so that when the users interact with the editor, things don't turn red too quickly
-        self.start_timer()
-
-    def update_background(self):
-        """
-            Update the markers based on which lines are detected as wrong
-        """
-        self.markerDeleteAll(0)
-        lines = self.wrong_format_lines
-        for line in lines:
-            self.markerAdd(line, 0)
-
-    def parse_content(self):
-        """
-            Parse the content of the editor (will be overridden by children classes)
-        """
-        return
-
-    def set_lexer(self):
-        """
-            Allow the user to edit the object
-        """
-        self.setLexer(self.lexer_)
-        self.setReadOnly(False)
-        self.is_lexed = True
-
-    def reinitialize(self):
-        """
-            Set the editor to its initial state (uneditable with empty background)
-        """
-        self.clear()
-        self.setLexer(None)
-        self.is_lexed = False
-        self.setReadOnly(True)
-        self.setPaper(self.empty_color)
-        self.markerDeleteAll()
-
-    def reset(self):
-        """
-            Clean the editor (i.e. remove content and reset attributes) but keep it editable
-        """
-        self.clear()
-        self.initial_content = {}
-        self.parse_and_format_editor()
-
-    @staticmethod
-    def to_format(input_string):
-        """
-            Turn the input string to the intended format (string, int, float or boolean)
-
-            @param input: String to convert
-            @return: Either a string, an int, a float or a boolean
-        """
-        try:
-            return int(input_string)
-        except (TypeError, ValueError):
-            pass
-        try:
-            return float(input_string)
-        except (TypeError, ValueError):
-            pass
-        if input_string in ("true", "True"):
-            return True
-        if input_string in ("false", "False"):
-            return False
-        return str(input_string)
-
-
-class YamlCodeEditor(GenericCodeEditor):
+class YamlCodeEditor(BaseCodeEditor):
 
     """
         QScintilla-based widget allowing to create a YAML code editor
@@ -269,18 +80,21 @@ class YamlCodeEditor(GenericCodeEditor):
         # If the editor has been emptied can just stop the parsing here
         if not editor_content:
             self.parsed_content = {}
-            self.contentIsModified.emit(self.initial_content != self.parsed_content)
+            self.contentIsModified.emit(
+                self.initial_content != self.parsed_content)
             return
         # Get the text line by line
         split_content = editor_content.split("\n")
         # Get indices of all lines starting without any space or \n (so lines defining a new root component)
-        zero_depth_indices = [i for i, x in enumerate(split_content) if re.search(r"^(\w+)", x) is not None]
+        zero_depth_indices = [i for i, x in enumerate(
+            split_content) if re.search(r"^(\w+)", x) is not None]
         # If nothing has been found (meaning that all the content is wrong), set all the lines in a single list
         if not zero_depth_indices:
             self.slices = [split_content]
         # If at least one is found, create slices from these indices (without forgetting from line 0 to first index)
         else:
-            self.slices = [split_content[:zero_depth_indices[0]]] if zero_depth_indices[0] else []
+            self.slices = [split_content[:zero_depth_indices[0]]
+                           ] if zero_depth_indices[0] else []
             self.slices += [split_content[zero_depth_indices[i]:zero_depth_indices[i + 1]]
                             for i in range(len(zero_depth_indices) - 1)] + [split_content[zero_depth_indices[-1]:]]
         # Current line number
@@ -303,7 +117,8 @@ class YamlCodeEditor(GenericCodeEditor):
                 # Search for spaces at the beginning of the line
                 starting_spaces = re.search(r"(^\s{1,})", line)
                 # Get the number of spaces out of the search
-                number_space = len(starting_spaces.group(1)) if starting_spaces else 0
+                number_space = len(starting_spaces.group(1)
+                                   ) if starting_spaces else 0
                 # As depth increases byt two spaces computes it
                 # To be valid the number of space should be even. If it's not the case set depth to a higher value
                 depth = number_space / 2 if number_space % 2 == 0 else expected_depth + 1
@@ -377,7 +192,8 @@ class YamlCodeEditor(GenericCodeEditor):
                 # Parse potential condensed dictionary
                 if condensed_dict:
                     content = re.search(r"\{(.*)\}", condensed_dict).group(1)
-                    dict_args = re.findall(r"([^\:\s\{\,]*)\s?:\s?([^\:\s\}\,]*)", condensed_dict)
+                    dict_args = re.findall(
+                        r"([^\:\s\{\,]*)\s?:\s?([^\:\s\}\,]*)", condensed_dict)
                     # If there is some text but not properly formatted mark the line as wrong and go to the next one
                     if not dict_args and content:
                         self.wrong_format_lines.append(line_number)
@@ -393,48 +209,58 @@ class YamlCodeEditor(GenericCodeEditor):
                         self.wrong_format_lines.append(line_number)
                         line_number += 1
                         continue
-                    formatted_dict_args = [(self.to_format(x), self.to_format(y)) for x, y in dict_args]
+                    formatted_dict_args = [
+                        (self.to_format(x), self.to_format(y)) for x, y in dict_args]
                     checked_condensed_dict = dict(formatted_dict_args)
 
                 # Parse potential condensed list
                 if condensed_list:
                     content = re.search(r"\[(.*)\]", condensed_list).group(1)
                     elements = content.split(",") if content else []
-                    checked_condensed_list = [self.to_format(element.strip()) for element in elements]
+                    checked_condensed_list = [self.to_format(
+                        element.strip()) for element in elements]
 
                 # Get the name of the parent component and corresponding line
-                parent_name = list(parent_dictionary[depth - 2])[-1] if depth else ""
+                parent_name = list(
+                    parent_dictionary[depth - 2])[-1] if depth else ""
                 parent_line = self._new_dicts_index[parent_name] if depth else -1
 
                 # If the line is at least composed of a keyname and column
                 if key_name and column:
                     # If a value is provided
                     if value:
-                        parent_dictionary[depth - 1][split_line[0]] = self.to_format(split_line[2])
+                        parent_dictionary[depth - 1][split_line[0]
+                                                     ] = self.to_format(split_line[2])
                     # If a condensed dictionary is provided
                     elif condensed_dict:
-                        parent_dictionary[depth - 1][split_line[0]] = checked_condensed_dict
+                        parent_dictionary[depth - 1][split_line[0]
+                                                     ] = checked_condensed_dict
                     elif condensed_list:
-                        parent_dictionary[depth - 1][split_line[0]] = checked_condensed_list
+                        parent_dictionary[depth - 1][split_line[0]
+                                                     ] = checked_condensed_list
                     # If no value is provided then create a new empty dictionary that is going to be filled by elements
                     # coming from higher depths
                     else:
                         new_empty_dict = {}
-                        parent_dictionary[depth - 1][split_line[0]] = new_empty_dict
+                        parent_dictionary[depth -
+                                          1][split_line[0]] = new_empty_dict
                         parent_dictionary[depth] = new_empty_dict
 
                 # If line corresponds to a list
                 elif dash:
                     # Get the object in which we should add the list
-                    object_to_fill = list(parent_dictionary[depth - 2].values())[-1]
+                    object_to_fill = list(
+                        parent_dictionary[depth - 2].values())[-1]
                     # If some text is provided after the dash
                     if list_element:
                         element = list_element.strip()
                         # look for a potential one element dictionary such as - a : b
-                        one_elem_dict = re.search(r"([^\:\s]*)\s?:\s?([^\:\s]*)", element)
+                        one_elem_dict = re.search(
+                            r"([^\:\s]*)\s?:\s?([^\:\s]*)", element)
                         # If the list_element starts with a space or the one element search found something incomplete
                         # make the line wrong and go the next one
-                        is_list_wrong = list_element.startswith(" ") or "," in list_element
+                        is_list_wrong = list_element.startswith(
+                            " ") or "," in list_element
                         if is_list_wrong or (one_elem_dict and not all(one_elem_dict.groups())):
                             self.wrong_format_lines.append(line_number)
                             line_number += 1
@@ -443,8 +269,10 @@ class YamlCodeEditor(GenericCodeEditor):
                         # Get either the simple element of the list or create a one element dictionary
                         if one_elem_dict:
                             key_, value_ = one_elem_dict.groups()
-                            formatted_elems = [(self.to_format(key_), self.to_format(value_))]
-                        element_to_add = dict(formatted_elems) if one_elem_dict else self.to_format(element)
+                            formatted_elems = [
+                                (self.to_format(key_), self.to_format(value_))]
+                        element_to_add = dict(
+                            formatted_elems) if one_elem_dict else self.to_format(element)
                     # If a condensed dict is provided
                     elif condensed_dict:
                         element_to_add = checked_condensed_dict
@@ -456,7 +284,8 @@ class YamlCodeEditor(GenericCodeEditor):
                         object_to_fill.append(element_to_add)
                     # Otherwise turns it to a list with the element inside
                     else:
-                        parent_dictionary[depth - 2][parent_name] = [element_to_add]
+                        parent_dictionary[depth -
+                                          2][parent_name] = [element_to_add]
                         # Delete the dictionary that was created in the parents before hand
                         del parent_dictionary[depth - 1]
 
@@ -467,13 +296,15 @@ class YamlCodeEditor(GenericCodeEditor):
         # Get the parsed content
         self.parsed_content = parsed
         # Emit the signal if it's different than the initial
-        self.contentIsModified.emit(self.initial_content != self.parsed_content)
+        self.contentIsModified.emit(
+            self.initial_content != self.parsed_content)
 
     def reset_init_content(self):
         """
             Reset the initial content
         """
-        self.initial_content = copy.deepcopy(self.parsed_content) if self.parsed_content else {}
+        self.initial_content = copy.deepcopy(
+            self.parsed_content) if self.parsed_content else {}
 
     def mark_component(self, component_name):
         """
@@ -488,7 +319,8 @@ class YamlCodeEditor(GenericCodeEditor):
         # Get the starting line
         begin_line = 0
         # If the element is part of new_dicts_index then get the beginning line from the dict
-        potential_key = self.slices[slice_index][0].strip(":").strip() if slice_index < len(self.slices) else -1
+        potential_key = self.slices[slice_index][0].strip(
+            ":").strip() if slice_index < len(self.slices) else -1
         if potential_key in self._new_dicts_index:
             begin_line = self._new_dicts_index[potential_key]
         # Otherwise it means it is a "complete" element and needs to get he beginning line
@@ -496,9 +328,11 @@ class YamlCodeEditor(GenericCodeEditor):
             for i in range(slice_index):
                 begin_line += len(self.slices[i])
         # Remove all trailing spaces of each element of the slice
-        striped_slice = list(map(lambda x: x.strip(), self.slices[slice_index]))
+        striped_slice = list(
+            map(lambda x: x.strip(), self.slices[slice_index]))
         # Get the ending line
-        end_line = begin_line + len(self.slices[slice_index]) - striped_slice.count("")
+        end_line = begin_line + \
+            len(self.slices[slice_index]) - striped_slice.count("")
         # For all these lines mark them as bad
         lines_indices = range(begin_line, end_line)
         for line_index in lines_indices:
@@ -527,7 +361,7 @@ class YamlCodeEditor(GenericCodeEditor):
         self.parsed_content = {}
 
 
-class XmlCodeEditor(GenericCodeEditor):
+class XmlCodeEditor(BaseCodeEditor):
 
     """
         QScintilla-based widget allowing to create a XML code editor
@@ -553,16 +387,20 @@ class XmlCodeEditor(GenericCodeEditor):
 
         if not editor_content:
             self.parsed_content = None
-            self.contentIsModified.emit(self.initial_content != self.parsed_content)
+            self.contentIsModified.emit(
+                self.initial_content != self.parsed_content)
             return
 
-        raw_arguments = re.search(r"\<include file=.*?\>(.*?)\<\/include\>", editor_content, re.DOTALL)
+        raw_arguments = re.search(
+            r"\<include file=.*?\>(.*?)\<\/include\>", editor_content, re.DOTALL)
         if raw_arguments is None:
             self.parsed_content = None
-            self.contentIsModified.emit(self.initial_content != self.parsed_content)
+            self.contentIsModified.emit(
+                self.initial_content != self.parsed_content)
             return
 
-        raw_arguments = re.sub("<!-- You can add any options you want to the file -->", "", raw_arguments.group(1))
+        raw_arguments = re.sub(
+            "<!-- You can add any options you want to the file -->", "", raw_arguments.group(1))
         # Strip is used to remove possible spaces at the head and tail of the string
         arguments_list = re.split("\n", raw_arguments.strip())
         filtered_arguments = [x.strip() for x in arguments_list if x]
@@ -573,13 +411,15 @@ class XmlCodeEditor(GenericCodeEditor):
         self.parsed_content = []
 
         for argument in filtered_arguments:
-            template_search = re.search(r"\<arg name\s?=\s?(.*?) value\s?=\s?(.*?)\s?\/\>", argument)
+            template_search = re.search(
+                r"\<arg name\s?=\s?(.*?) value\s?=\s?(.*?)\s?\/\>", argument)
             if template_search is None:
                 self.wrong_format_lines.append(filtered_editor.index(argument))
             else:
                 self.parsed_content.append(argument)
 
-        self.contentIsModified.emit(self.initial_content != self.parsed_content)
+        self.contentIsModified.emit(
+            self.initial_content != self.parsed_content)
 
     def update_background(self):
         """
